indexer moved to mod piop
