+++ ring/src/lib.rs
@@ -1,97 +1,16 @@
-use std::marker::PhantomData;
+use fflonk::pcs::PCS;
 
-use ark_ec::AffineRepr;
-use ark_ec::short_weierstrass::{Affine, SWCurveConfig};
-use ark_ff::PrimeField;
-use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
-use fflonk::pcs::{Commitment, PCS, PcsParams};
-
-use common::{Column, FieldColumn, Proof};
-use common::gadgets::sw_cond_add::AffineColumn;
+use common::Proof;
 
 use crate::piop::{RingCommitments, RingEvaluations};
 use crate::piop::params::PiopParams;
 
 mod piop;
+pub use piop::index;
 pub mod ring_prover;
 pub mod ring_verifier;
 
 type RingProof<F, CS> = Proof<F, CS, RingCommitments<F, <CS as PCS<F>>::C>, RingEvaluations<F>>;
-
-// Columns commitment to which the verifier knows (or trusts).
-#[derive(Clone)]
-pub struct FixedColumns<F: PrimeField, G: AffineRepr<BaseField=F>> {
-    // Public keys of the ring participants in order,
-    // followed by the powers-of-2 multiples of the second Pedersen base.
-    // pk_1, ..., pk_n, H, 2H, 4H, ..., 2^sH
-    // 1          n                     n+s+1
-    points: AffineColumn<F, G>,
-    // Binary column that highlights which rows of the table correspond to the ring.
-    // 1, 1, ..., 1, 0, 0, ..., 0
-    // 1          n
-    ring_selector: FieldColumn<F>,
-}
-
-// Commitments to the fixed columns (see above).
-#[derive(Clone, CanonicalSerialize, CanonicalDeserialize)]
-pub struct FixedColumnsCommitted<F: PrimeField, C: Commitment<F>> {
-    points: [C; 2],
-    ring_selector: C,
-    phantom: PhantomData<F>,
-}
-
-impl<F: PrimeField, C: Commitment<F>> FixedColumnsCommitted<F, C> {
-    fn as_vec(&self) -> Vec<C> {
-        vec![
-            self.points[0].clone(),
-            self.points[1].clone(),
-            self.ring_selector.clone(),
-        ]
-    }
-}
-
-impl<F: PrimeField, G: AffineRepr<BaseField=F>> FixedColumns<F, G> {
-    fn commit<CS: PCS<F>>(&self, ck: &CS::CK) -> FixedColumnsCommitted<F, CS::C> {
-        let points = [
-            CS::commit(ck, self.points.xs.as_poly()),
-            CS::commit(ck, self.points.ys.as_poly()),
-        ];
-        let ring_selector = CS::commit(ck, self.ring_selector.as_poly());
-        FixedColumnsCommitted { points, ring_selector, phantom: Default::default() }
-    }
-}
-
-pub struct ProverKey<F: PrimeField, CS: PCS<F>, G: AffineRepr<BaseField=F>> {
-    pcs_ck: CS::CK,
-    fixed_columns: FixedColumns<F, G>,
-    verifier_key: VerifierKey<F, CS>, // used in the Fiat-Shamir transform
-}
-
-
-#[derive(Clone, CanonicalSerialize, CanonicalDeserialize)]
-pub struct VerifierKey<F: PrimeField, CS: PCS<F>> {
-    pcs_raw_vk: <CS::Params as PcsParams>::RVK,
-    fixed_columns_committed: FixedColumnsCommitted<F, CS::C>,
-    //TODO: domain
-}
-
-pub fn index<F: PrimeField, CS: PCS<F>, Curve: SWCurveConfig<BaseField=F>>(
-    pcs_params: CS::Params,
-    piop_params: &PiopParams<F, Curve>,
-    keys: Vec<Affine<Curve>>,
-) -> (ProverKey<F, CS, Affine<Curve>>, VerifierKey<F, CS>) {
-    let pcs_ck = pcs_params.ck();
-    let pcs_raw_vk = pcs_params.raw_vk();
-    let fixed_columns = piop_params.fixed_columns(&keys);
-    let fixed_columns_committed = fixed_columns.commit::<CS>(&pcs_ck);
-    let verifier_key = VerifierKey {
-        pcs_raw_vk: pcs_raw_vk.clone(),
-        fixed_columns_committed: fixed_columns_committed.clone(),
-    };
-    let prover_key = ProverKey { pcs_ck, fixed_columns, verifier_key };
-    let verifier_key = VerifierKey { pcs_raw_vk, fixed_columns_committed };
-    (prover_key, verifier_key)
-}
 
 #[cfg(test)]
 mod tests {
@@ -127,7 +46,7 @@ mod tests {
         let k = rng.gen_range(0..keyset_size); // prover's secret index
         let pk = pks[k].clone();
 
-        let (prover_key, verifier_key) = crate::index::<_, CS, _>(pcs_params, &piop_params, pks);
+        let (prover_key, verifier_key) = crate::piop::index::<_, CS, _>(pcs_params, &piop_params, pks);
 
         // PROOF generation
         let secret = Fr::rand(rng); // prover's secret scalar
