+++ ring/src/piop/prover.rs
@@ -14,7 +14,7 @@ use common::gadgets::inner_prod::InnerProd;
 use common::gadgets::ProverGadget;
 use common::gadgets::sw_cond_add::{AffineColumn, CondAdd};
 use common::piop::ProverPiop;
-use crate::FixedColumns;
+use crate::piop::FixedColumns;
 
 use crate::piop::{RingCommitments, RingEvaluations};
 use crate::piop::params::PiopParams;
